Module Channel.Transaction

Transactions. Transactions can be made per channel.

After a transaction is started, all published messages and all changes (queue/exchange bindings, creations or deletions) and message acknowledgements are not visible outside the transaction.

The changes becomes visible after a commit or canceled by call to rollback.

type tx

val start : _ t -> tx Thread.Deferred.t

Start a transacction

val commit : tx -> unit Thread.Deferred.t

Commit an transaction

val rollback : tx -> unit Thread.Deferred.t

Rollback a transaction, discarding all changes and messages