Module Amqp_client_async.Constants

Internal - Low level protocol description

val frame_method : int

val frame_header : int

val frame_body : int

val frame_heartbeat : int

val frame_min_size : int

val frame_end : int

val reply_success : int

Indicates that the method completed successfully. This reply code is reserved for future use - the current protocol design does not use positive confirmation and reply codes are sent only in case of an error.

val content_too_large : int

The client attempted to transfer content larger than the server could accept at the present time. The client may retry at a later time.

val no_route : int

Returned when RabbitMQ sends back with 'basic.return' when a 'mandatory' message cannot be delivered to any queue.

val no_consumers : int

When the exchange cannot deliver to a consumer when the immediate flag is set. As a result of pending data on the queue or the absence of any consumers of the queue.

val connection_forced : int

An operator intervened to close the connection for some reason. The client may retry at some later date.

val invalid_path : int

The client tried to work with an unknown virtual host.

val access_refused : int

The client attempted to work with a server entity to which it has no access due to security settings.

val not_found : int

The client attempted to work with a server entity that does not exist.

val resource_locked : int

The client attempted to work with a server entity to which it has no access because another client is working with it.

val precondition_failed : int

The client requested a method that was not allowed because some precondition failed.

val frame_error : int

The sender sent a malformed frame that the recipient could not decode. This strongly implies a programming error in the sending peer.

val syntax_error : int

The sender sent a frame that contained illegal values for one or more fields. This strongly implies a programming error in the sending peer.

val command_invalid : int

The client sent an invalid sequence of frames, attempting to perform an operation that was considered invalid by the server. This usually implies a programming error in the client.

val channel_error : int

The client attempted to work with a channel that had not been correctly opened. This most likely indicates a fault in the client layer.

val unexpected_frame : int

The peer sent a frame that was not expected, usually in the context of a content header and body. This strongly indicates a fault in the peer's content processing.

val resource_error : int

The server could not complete the method because it lacked sufficient resources. This may be due to the client creating too many of some type of entity.

val not_allowed : int

The client tried to work with some entity in a manner that is prohibited by the server, due to security settings or by some other criteria.

val not_implemented : int

The client tried to use functionality that is not implemented in the server.

val internal_error : int

The server could not complete the method because of an internal error. The server may require intervention by an operator in order to resume normal operations.