Module Spec.Basic

The Basic class provides methods that support an industry-standard messaging model.

module Content : sig ... end

module Qos_ok : sig ... end

This method tells the client that the requested QoS levels could be handled by the server. The requested QoS applies to all active consumers until a new QoS is defined.

module Qos : sig ... end

This method requests a specific quality of service. The QoS can be specified for the current channel or for all channels on the connection. The particular properties and semantics of a qos method always depend on the content class semantics. Though the qos method could in principle apply to both peers, it is currently meaningful only for the server.

module Consume_ok : sig ... end

The server provides the client with a consumer tag, which is used by the client for methods called on the consumer at a later stage.

module Consume : sig ... end

This method asks the server to start a "consumer", which is a transient request for messages from a specific queue. Consumers last as long as the channel they were declared on, or until the client cancels them.

module Cancel_ok : sig ... end

This method confirms that the cancellation was completed.

module Cancel : sig ... end

This method cancels a consumer. This does not affect already delivered messages, but it does mean the server will not send any more messages for that consumer. The client may receive an arbitrary number of messages in between sending the cancel method and receiving the cancel-ok reply.

module Publish : sig ... end

This method publishes a message to a specific exchange. The message will be routed to queues as defined by the exchange configuration and distributed to any active consumers when the transaction, if any, is committed.

module Return : sig ... end

This method returns an undeliverable message that was published with the "immediate" flag set, or an unroutable message published with the "mandatory" flag set. The reply code and text provide information about the reason that the message was undeliverable.

module Deliver : sig ... end

This method delivers a message to the client, via a consumer. In the asynchronous message delivery model, the client starts a consumer using the Consume method, then the server responds with Deliver methods as and when messages arrive for that consumer.

module Get_ok : sig ... end

This method delivers a message to the client following a get method. A message delivered by 'get-ok' must be acknowledged unless the no-ack option was set in the get method.

module Get_empty : sig ... end

This method tells the client that the queue has no messages available for the client.

module Get : sig ... end

This method provides a direct access to the messages in a queue using a synchronous dialogue that is designed for specific types of application where synchronous functionality is more important than performance.

module Ack : sig ... end

When sent by the client, this method acknowledges one or more messages delivered via the Deliver or Get-Ok methods.

module Reject : sig ... end

This method allows a client to reject a message. It can be used to interrupt and cancel large incoming messages, or return untreatable messages to their original queue.

module Recover_async : sig ... end

This method asks the server to redeliver all unacknowledged messages on a specified channel. Zero or more messages may be redelivered. This method is deprecated in favour of the synchronous Recover/Recover-Ok.

module Recover : sig ... end

This method asks the server to redeliver all unacknowledged messages on a specified channel. Zero or more messages may be redelivered. This method replaces the asynchronous Recover.

module Recover_ok : sig ... end

This method acknowledges a Basic.Recover method.

module Nack : sig ... end

This method allows a client to reject one or more incoming messages. It can be used to interrupt and cancel large incoming messages, or return untreatable messages to their original queue.