Module Basic.Consume_ok

The server provides the client with a consumer tag, which is used by the client for methods called on the consumer at a later stage.

type t = {

consumer_tag : consumer_tag;
(*
Holds the consumer tag specified by the client or provided by the server.
*)

}

val init : consumer_tag:consumer_tag -> unit -> t