Module Basic.Deliver

This method delivers a message to the client, via a consumer. In the asynchronous message delivery model, the client starts a consumer using the Consume method, then the server responds with Deliver methods as and when messages arrive for that consumer.

type t = {

consumer_tag : consumer_tag;

delivery_tag : delivery_tag;

redelivered : redelivered;

exchange : exchange_name;
(*
Specifies the name of the exchange that the message was originally published to. May be empty, indicating the default exchange.
*)

routing_key : Amqp_client_lib.Types.shortstr;
(*
Specifies the routing key name specified when the message was published.
*)

}

val init : consumer_tag:consumer_tag -> delivery_tag:delivery_tag -> redelivered:redelivered -> exchange:exchange_name -> routing_key:Amqp_client_lib.Types.shortstr -> unit -> t

val reply : ?once:bool -> (Framing.t * Framing.channel_no) -> (t * (Content.t * string)) Async_kernel.Deferred.t