Module Basic.Get

This method provides a direct access to the messages in a queue using a synchronous dialogue that is designed for specific types of application where synchronous functionality is more important than performance.

type t = {

queue : queue_name;
(*
Specifies the name of the queue to get a message from.
*)

no_ack : no_ack;

}

val init : queue:'a -> no_ack:queue_name -> unit -> no_ack -> t

val request : (Framing.t * Framing.channel_no) -> t -> [> `Get_empty of unit | `Get_ok of Get_ok.t * (Content.t * string) ] Thread.Deferred.t