Module Basic.Qos_ok

This method tells the client that the requested QoS levels could be handled by the server. The requested QoS applies to all active consumers until a new QoS is defined.

type t = unit

val init : unit -> unit