Module Basic.Recover_async

This method asks the server to redeliver all unacknowledged messages on a specified channel. Zero or more messages may be redelivered. This method is deprecated in favour of the synchronous Recover/Recover-Ok.

type t = {

requeue : Amqp_client_lib.Types.bit;
(*
If this field is zero, the message will be redelivered to the original recipient. If this bit is 1, the server will attempt to requeue the message, potentially then delivering it to an alternative subscriber.
*)

}

val init : requeue:Amqp_client_lib.Types.bit -> unit -> t

val request : (Framing.t * Framing.channel_no) -> t -> unit Thread.Deferred.t