Module Basic.Reject

This method allows a client to reject a message. It can be used to interrupt and cancel large incoming messages, or return untreatable messages to their original queue.

type t = {

delivery_tag : delivery_tag;

requeue : Amqp_client_lib.Types.bit;
(*
If requeue is true, the server will attempt to requeue the message. If requeue is false or the requeue attempt fails the messages are discarded or dead-lettered.
*)

}

val init : delivery_tag:delivery_tag -> requeue:Amqp_client_lib.Types.bit -> unit -> t

val request : (Framing.t * Framing.channel_no) -> t -> unit Thread.Deferred.t