Module Spec.Channel

The channel class provides methods for a client to establish a channel to a server and for both peers to operate the channel thereafter.

module Open_ok : sig ... end

This method signals to the client that the channel is ready for use.

module Open : sig ... end

This method opens a channel to the server.

module Flow_ok : sig ... end

Confirms to the peer that a flow command was received and processed.

module Flow : sig ... end

This method asks the peer to pause or restart the flow of content data sent by a consumer. This is a simple flow-control mechanism that a peer can use to avoid overflowing its queues or otherwise finding itself receiving more messages than it can process. Note that this method is not intended for window control. It does not affect contents returned by Basic.Get-Ok methods.

module Close_ok : sig ... end

This method confirms a Channel.Close method and tells the recipient that it is safe to release resources for the channel.

module Close : sig ... end

This method indicates that the sender wants to close the channel. This may be due to internal conditions (e.g. a forced shut-down) or due to an error handling a specific method, i.e. an exception. When a close is due to an exception, the sender provides the class and method id of the method which caused the exception.