Module Confirm.Select

This method sets the channel to use publisher acknowledgements. The client can only use this method on a non-transactional channel.

type t = {

nowait : no_wait;

}

val init : nowait:no_wait -> unit -> t

val request : (Framing.t * Framing.channel_no) -> t -> unit Thread.Deferred.t