Module Connection.Secure_ok

This method attempts to authenticate, passing a block of SASL data for the security mechanism at the server side.

type t = {

response : Amqp_client_lib.Types.longstr;
A block of opaque data passed to the security mechanism. The contents of this data are defined by the SASL security mechanism.

}

val init : response:Amqp_client_lib.Types.longstr -> unit -> t