Module Connection.Start

This method starts the connection negotiation process by telling the client the protocol version that the server proposes, along with a list of security mechanisms which the client can use for authentication.

type t = {

version_major : Amqp_client_lib.Types.octet;
(*
The major version number can take any value from 0 to 99 as defined in the AMQP specification.
*)

version_minor : Amqp_client_lib.Types.octet;
(*
The minor version number can take any value from 0 to 99 as defined in the AMQP specification.
*)

server_properties : peer_properties;

mechanisms : Amqp_client_lib.Types.longstr;
(*
A list of the security mechanisms that the server supports, delimited by spaces.
*)

locales : Amqp_client_lib.Types.longstr;
(*
A list of the message locales that the server supports, delimited by spaces. The locale defines the language in which the server will send reply texts.
*)

}

val init : version_major:Amqp_client_lib.Types.octet -> version_minor:Amqp_client_lib.Types.octet -> server_properties:peer_properties -> mechanisms:Amqp_client_lib.Types.longstr -> locales:Amqp_client_lib.Types.longstr -> unit -> t

val reply : ?once:bool -> (Framing.t * Framing.channel_no) -> (t -> Start_ok.t Thread.Deferred.t) -> unit Thread.Deferred.t