Module Connection.Start_ok

This method selects a SASL security mechanism.

type t = {

client_properties : peer_properties;

mechanism : Amqp_client_lib.Types.shortstr;
(*
A single security mechanisms selected by the client, which must be one of those specified by the server.
*)

response : Amqp_client_lib.Types.longstr;
(*
A block of opaque data passed to the security mechanism. The contents of this data are defined by the SASL security mechanism.
*)

locale : Amqp_client_lib.Types.shortstr;
(*
A single message locale selected by the client, which must be one of those specified by the server.
*)

}

val init : client_properties:peer_properties -> mechanism:Amqp_client_lib.Types.shortstr -> response:Amqp_client_lib.Types.longstr -> locale:Amqp_client_lib.Types.shortstr -> unit -> t