Module Connection.Tune_ok

This method sends the client's connection tuning parameters to the server. Certain fields are negotiated, others provide capability information.

type t = {

channel_max : Amqp_client_lib.Types.short;
(*
The maximum total number of channels that the client will use per connection.
*)

frame_max : Amqp_client_lib.Types.long;
(*
The largest frame size that the client and server will use for the connection. Zero means that the client does not impose any specific limit but may reject very large frames if it cannot allocate resources for them. Note that the frame-max limit applies principally to content frames, where large contents can be broken into frames of arbitrary size.
*)

heartbeat : Amqp_client_lib.Types.short;
(*
The delay, in seconds, of the connection heartbeat that the client wants. Zero means the client does not want a heartbeat.
*)

}

val init : channel_max:Amqp_client_lib.Types.short -> frame_max:Amqp_client_lib.Types.long -> heartbeat:Amqp_client_lib.Types.short -> unit -> t