Module Spec.Exchange

Exchanges match and distribute messages across queues. Exchanges can be configured in the server or declared at runtime.

module Declare_ok : sig ... end

This method confirms a Declare method and confirms the name of the exchange, essential for automatically-named exchanges.

module Declare : sig ... end

This method creates an exchange if it does not already exist, and if the exchange exists, verifies that it is of the correct and expected class.

module Delete_ok : sig ... end

This method confirms the deletion of an exchange.

module Delete : sig ... end

This method deletes an exchange. When an exchange is deleted all queue bindings on the exchange are cancelled.

module Bind_ok : sig ... end

This method confirms that the bind was successful.

module Bind : sig ... end

This method binds an exchange to an exchange.

module Unbind_ok : sig ... end

This method confirms that the unbind was successful.

module Unbind : sig ... end

This method unbinds an exchange from an exchange.