Module Exchange.Delete

This method deletes an exchange. When an exchange is deleted all queue bindings on the exchange are cancelled.

type t = {

exchange : exchange_name;

if_unused : Amqp_client_lib.Types.bit;
(*
If set, the server will only delete the exchange if it has no queue bindings. If the exchange has queue bindings the server does not delete it but raises a channel exception instead.
*)

no_wait : no_wait;

}

val init : exchange:'a -> if_unused:exchange_name -> no_wait:Amqp_client_lib.Types.bit -> unit -> no_wait -> t

val request : (Framing.t * Framing.channel_no) -> t -> unit Thread.Deferred.t