Module Queue.Declare_ok

This method confirms a Declare method and confirms the name of the queue, essential for automatically-named queues.

type t = {

queue : queue_name;
(*
Reports the name of the queue. If the server generated a queue name, this field contains that name.
*)

message_count : message_count;

consumer_count : Amqp_client_lib.Types.long;
(*
Reports the number of active consumers for the queue. Note that consumers can suspend activity (Channel.Flow) in which case they do not appear in this count.
*)

}

val init : queue:queue_name -> message_count:message_count -> consumer_count:Amqp_client_lib.Types.long -> unit -> t