Module Queue.Delete

This method deletes a queue. When a queue is deleted any pending messages are sent to a dead-letter queue if this is defined in the server configuration, and all consumers on the queue are cancelled.

type t = {

queue : queue_name;
(*
Specifies the name of the queue to delete.
*)

if_unused : Amqp_client_lib.Types.bit;
(*
If set, the server will only delete the queue if it has no consumers. If the queue has consumers the server does does not delete it but raises a channel exception instead.
*)

if_empty : Amqp_client_lib.Types.bit;
(*
If set, the server will only delete the queue if it has no messages.
*)

no_wait : no_wait;

}

val init : queue:'a -> if_unused:queue_name -> if_empty:Amqp_client_lib.Types.bit -> no_wait:Amqp_client_lib.Types.bit -> unit -> no_wait -> t

val request : (Framing.t * Framing.channel_no) -> t -> Delete_ok.t Thread.Deferred.t