Module Tx.Rollback

This method abandons all message publications and acknowledgments performed in the current transaction. A new transaction starts immediately after a rollback. Note that unacked messages will not be automatically redelivered by rollback; if that is required an explicit recover call should be issued.

type t = unit

val init : unit -> unit

val request : (Framing.t * Framing.channel_no) -> unit -> unit Thread.Deferred.t