Module Tx.Select

This method sets the channel to use standard transactions. The client must use this method at least once on a channel before using the Commit or Rollback methods.

type t = unit

val init : unit -> unit

val request : (Framing.t * Framing.channel_no) -> unit -> unit Thread.Deferred.t